add comments to util file
